+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## [v.0.5.0] - 2025/03/07
+
+### Added
+- WalletCore framework
+
+### Removed
+- x86_64 architecture support
+
 ## [v.0.2.0] - 2025/03/06
 
 ### Added